+++ lib/pocket-ruby.rb
@@ -3,6 +3,7 @@ require File.expand_path("../pocket/configuration", __FILE__)
 require File.expand_path("../pocket/api", __FILE__)
 require File.expand_path("../pocket/client", __FILE__)
 require File.expand_path("../pocket/article", __FILE__)
+require File.expand_path("../pocket/author", __FILE__)
 
 module Pocket
   extend Configuration